Update Items.py with latest data from mafia.
In doing so, add a number of new enchantments to ItemsSerializer.
Also upgrade ItemDataConverter so we are able to merge new changes in
from mafia to our current data files. To make sure that new items
always get the correct 'type', when we find a new item in mafia we also
login to KoL and grab all of the information we can about that item.