Fix code to work with recent mall changes.
Mall searches and purchases should now work again. This was only a basic
fix and some functionality has been removed. For example, the code does
not determine if the user has already hit the limit in a particular
store. In addition, the mall search no longer returns items that are not
in the item database. These features should be added back in.