Fix auto-discovery of new items.
The ClosetContentsRequest was broken by a recent server-side change. We
now use the API to get the user's inventory and then loop over all
items, looking for ones that are not known. For each unknown item, we
ask the server what it is and then add it to the ItemDatabase.